Re-arrange files
